CXXCBC-489: Support scoped eventing functions
