+++ core/operations/management/eventing_get_function.cxx
@@ -18,6 +18,7 @@
 #include "eventing_get_function.hxx"
 #include "core/management/eventing_function_json.hxx"
 #include "core/utils/json.hxx"
+#include "core/utils/url_codec.hxx"
 #include "error_utils.hxx"
 
 #include <fmt/core.h>
@@ -29,6 +30,11 @@ eventing_get_function_request::encode_to(encoded_request_type& encoded, http_con
 {
     encoded.method = "GET";
     encoded.path = fmt::format("/api/v1/functions/{}", name);
+    if (bucket_name.has_value() && scope_name.has_value()) {
+        encoded.path += fmt::format("?bucket={}&scope={}",
+                                    utils::string_codec::v2::path_escape(bucket_name.value()),
+                                    utils::string_codec::v2::path_escape(scope_name.value()));
+    }
     return {};
 }
