+++ core/operations/management/eventing_pause_function.cxx
@@ -16,7 +16,9 @@
  */
 
 #include "eventing_pause_function.hxx"
+
 #include "core/utils/json.hxx"
+#include "core/utils/url_codec.hxx"
 #include "error_utils.hxx"
 
 #include <fmt/core.h>
@@ -28,6 +30,11 @@ eventing_pause_function_request::encode_to(encoded_request_type& encoded, http_c
 {
     encoded.method = "POST";
     encoded.path = fmt::format("/api/v1/functions/{}/pause", name);
+    if (bucket_name.has_value() && scope_name.has_value()) {
+        encoded.path += fmt::format("?bucket={}&scope={}",
+                                    utils::string_codec::v2::path_escape(bucket_name.value()),
+                                    utils::string_codec::v2::path_escape(scope_name.value()));
+    }
     return {};
 }
