+++ core/operations/management/eventing_upsert_function.hxx
@@ -40,6 +40,9 @@ struct eventing_upsert_function_request {
 
     couchbase::core::management::eventing::function function{};
 
+    std::optional<std::string> bucket_name{};
+    std::optional<std::string> scope_name{};
+
     static const inline service_type type = service_type::eventing;
 
     std::optional<std::string> client_context_id{};
