+++ test/test_integration_management.cxx
@@ -21,16 +21,9 @@
 #include <catch2/catch_approx.hpp>
 
 #include "core/management/analytics_link.hxx"
-#include "core/operations/document_append.hxx"
-#include "core/operations/document_decrement.hxx"
 #include "core/operations/document_get.hxx"
-#include "core/operations/document_increment.hxx"
 #include "core/operations/document_insert.hxx"
 #include "core/operations/document_mutate_in.hxx"
-#include "core/operations/document_prepend.hxx"
-#include "core/operations/document_remove.hxx"
-#include "core/operations/document_replace.hxx"
-#include "core/operations/document_upsert.hxx"
 #include "core/operations/management/analytics.hxx"
 #include "core/operations/management/bucket.hxx"
 #include "core/operations/management/collections.hxx"
@@ -5097,246 +5090,5 @@ TEST_CASE("integration: freeform HTTP request", "[integration]")
         auto result = couchbase::core::utils::json::parse(resp.body);
         INFO(resp.body);
         REQUIRE(result.is_array());
-    }
-}
-
-static couchbase::core::operations::management::eventing_get_function_response
-wait_for_function_created(test::utils::integration_test_guard& integration, const std::string& function_name)
-{
-    couchbase::core::operations::management::eventing_get_function_response resp{};
-    test::utils::wait_until([&integration, &resp, function_name]() {
-        couchbase::core::operations::management::eventing_get_function_request req{ function_name };
-        resp = test::utils::execute(integration.cluster, req);
-        return !resp.ctx.ec;
-    });
-    return resp;
-}
-
-static bool
-wait_for_function_reach_status(test::utils::integration_test_guard& integration,
-                               const std::string& function_name,
-                               couchbase::core::management::eventing::function_status status)
-{
-    return test::utils::wait_until(
-      [&integration, function_name, status]() {
-          couchbase::core::operations::management::eventing_get_status_request req{};
-          auto resp = test::utils::execute(integration.cluster, req);
-          if (resp.ctx.ec) {
-              return false;
-          }
-          auto function = std::find_if(resp.status.functions.begin(), resp.status.functions.end(), [function_name](const auto& fun) {
-              return function_name == fun.name;
-          });
-          if (function == resp.status.functions.end()) {
-              return false;
-          }
-          return function->status == status;
-      },
-      std::chrono::minutes(3));
-}
-
-TEST_CASE("integration: eventing functions management", "[integration]")
-{
-    test::utils::integration_test_guard integration;
-
-    if (!integration.cluster_version().supports_eventing_functions()) {
-        SKIP("cluster does not support eventing service");
-    }
-    if (!integration.has_eventing_service()) {
-        SKIP("cluster does not have eventing service");
-    }
-
-    if (!integration.cluster_version().supports_gcccp()) {
-        test::utils::open_bucket(integration.cluster, integration.ctx.bucket);
-    }
-
-    SECTION("lifecycle")
-    {
-        auto function_name = test::utils::uniq_id("name");
-
-        {
-            couchbase::core::operations::management::eventing_drop_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            if (integration.cluster_version().is_cheshire_cat()) {
-                REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_not_deployed);
-            } else {
-                REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_not_found);
-            }
-        }
-
-        {
-            couchbase::core::operations::management::eventing_get_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_not_found);
-        }
-
-        auto meta_bucket_name = test::utils::uniq_id("meta");
-        {
-
-            couchbase::core::management::cluster::bucket_settings bucket_settings;
-            bucket_settings.name = meta_bucket_name;
-            bucket_settings.ram_quota_mb = 256;
-
-            {
-                couchbase::core::operations::management::bucket_create_request req;
-                req.bucket = bucket_settings;
-                auto resp = test::utils::execute(integration.cluster, req);
-                REQUIRE_SUCCESS(resp.ctx.ec);
-            }
-        }
-
-        {
-            auto resp = wait_for_bucket_created(integration, meta_bucket_name);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        std::string source_code = R"(
-function OnUpdate(doc, meta) {
-    log("Doc created/updated", meta.id);
-}
-
-function OnDelete(meta, options) {
-    log("Doc deleted/expired", meta.id);
-}
-)";
-
-        {
-            couchbase::core::operations::management::eventing_upsert_function_request req{};
-            req.function.source_keyspace.bucket = integration.ctx.bucket;
-            req.function.metadata_keyspace.bucket = meta_bucket_name;
-            req.function.name = function_name;
-            req.function.code = source_code;
-            req.function.settings.handler_headers = { "// generated by Couchbase C++ SDK" };
-            req.function.constant_bindings.emplace_back(couchbase::core::management::eventing::function_constant_binding{ "PI", "3.14" });
-            req.function.bucket_bindings.emplace_back(couchbase::core::management::eventing::function_bucket_binding{
-              "data", { integration.ctx.bucket }, couchbase::core::management::eventing::function_bucket_access::read_write });
-            req.function.url_bindings.emplace_back(
-              couchbase::core::management::eventing::function_url_binding{ "home", "https://couchbase.com" });
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        {
-            auto resp = wait_for_function_created(integration, function_name);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_get_all_functions_request req{};
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-            auto function = std::find_if(
-              resp.functions.begin(), resp.functions.end(), [&function_name](const auto& fun) { return function_name == fun.name; });
-            REQUIRE(function != resp.functions.end());
-            REQUIRE(function->code == source_code);
-            REQUIRE(function->source_keyspace.bucket == integration.ctx.bucket);
-            REQUIRE(function->metadata_keyspace.bucket == meta_bucket_name);
-            REQUIRE(function->settings.deployment_status == couchbase::core::management::eventing::function_deployment_status::undeployed);
-            REQUIRE(function->settings.processing_status == couchbase::core::management::eventing::function_processing_status::paused);
-            REQUIRE(!function->settings.handler_headers.empty());
-            REQUIRE(function->settings.handler_headers[0] == "// generated by Couchbase C++ SDK");
-            REQUIRE(!function->constant_bindings.empty());
-            REQUIRE(function->constant_bindings[0].alias == "PI");
-            REQUIRE(function->constant_bindings[0].literal == "3.14");
-            REQUIRE(!function->bucket_bindings.empty());
-            REQUIRE(function->bucket_bindings[0].alias == "data");
-            REQUIRE(function->bucket_bindings[0].name.bucket == "default");
-            REQUIRE(function->bucket_bindings[0].access == couchbase::core::management::eventing::function_bucket_access::read_write);
-            REQUIRE(!function->url_bindings.empty());
-            REQUIRE(function->url_bindings[0].alias == "home");
-            REQUIRE(function->url_bindings[0].hostname == "https://couchbase.com");
-            REQUIRE(std::holds_alternative<couchbase::core::management::eventing::function_url_no_auth>(function->url_bindings[0].auth));
-        }
-
-        {
-            couchbase::core::operations::management::eventing_get_status_request req{};
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-            REQUIRE(resp.status.num_eventing_nodes > 0);
-            auto function = std::find_if(resp.status.functions.begin(), resp.status.functions.end(), [&function_name](const auto& fun) {
-                return function_name == fun.name;
-            });
-            REQUIRE(function != resp.status.functions.end());
-            REQUIRE(function->status == couchbase::core::management::eventing::function_status::undeployed);
-            REQUIRE(function->deployment_status == couchbase::core::management::eventing::function_deployment_status::undeployed);
-            REQUIRE(function->processing_status == couchbase::core::management::eventing::function_processing_status::paused);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_undeploy_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_not_deployed);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_deploy_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        REQUIRE(
-          wait_for_function_reach_status(integration, function_name, couchbase::core::management::eventing::function_status::deployed));
-
-        {
-            couchbase::core::operations::management::eventing_drop_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_deployed);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_resume_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_deployed);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_pause_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        REQUIRE(wait_for_function_reach_status(integration, function_name, couchbase::core::management::eventing::function_status::paused));
-
-        {
-            couchbase::core::operations::management::eventing_pause_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_paused);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_resume_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        REQUIRE(
-          wait_for_function_reach_status(integration, function_name, couchbase::core::management::eventing::function_status::deployed));
-
-        {
-            couchbase::core::operations::management::eventing_undeploy_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        REQUIRE(
-          wait_for_function_reach_status(integration, function_name, couchbase::core::management::eventing::function_status::undeployed));
-
-        {
-            couchbase::core::operations::management::eventing_drop_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
-
-        {
-            couchbase::core::operations::management::eventing_get_function_request req{ function_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE(resp.ctx.ec == couchbase::errc::management::eventing_function_not_found);
-        }
-
-        {
-            couchbase::core::operations::management::bucket_drop_request req{ meta_bucket_name };
-            auto resp = test::utils::execute(integration.cluster, req);
-            REQUIRE_SUCCESS(resp.ctx.ec);
-        }
     }
 }
