+++ test/utils/server_version.hxx
@@ -176,6 +176,11 @@ struct server_version {
         return !use_gocaves && is_enterprise() && (is_cheshire_cat() || is_neo()) && deployment == deployment_type::on_prem;
     }
 
+    [[nodiscard]] bool supports_scoped_eventing_functions() const
+    {
+        return !use_gocaves && is_enterprise() && is_neo() && deployment == deployment_type::on_prem;
+    }
+
     [[nodiscard]] bool supports_scope_search() const
     {
         return (major == 7 && minor >= 6) || major > 7;
